+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - M68k Browser: 
   - re-use default IDE documentation component for better UX
   - various smaller layout/content improvements
+  - "Amiga Hardware" tab: add _Related_ registers
 
 ### Added